graphql restructure
