+++ sqrl/graphql/schemas/timetable.py
@@ -1,171 +1,25 @@
-from typing import Any, Optional
+"""Timetable queries and mutations."""
+from typing import Optional
 
 import graphene
-import mongoengine
-from flask import current_app
-from graphene_mongo import MongoengineObjectType
 
-from sqrl.models.common import Time
-from sqrl.models.timetable import (Course, Instructor, Organisation, Section,
-                                   SectionMeeting, Timetable)
+from gator.models.timetable import Course
+
+from sqrl.graphql.objects.timetable import UserTimetableObject
+from sqrl.models import UserTimetable
 
 
-class _TimeObject(MongoengineObjectType):
-    """A time object in the graphql schema."""
-
-    class Meta:
-        model = Time
-
-
-class _SectionMeetingObject(MongoengineObjectType):
-    """A section meeting in the graphql schema."""
-
-    class Meta:
-        model = SectionMeeting
-
-
-class _InstructorObject(MongoengineObjectType):
-    """An instructor in the graphql schema."""
-
-    class Meta:
-        model = Instructor
-
-
-class _SectionObject(MongoengineObjectType):
-    """A section in the graphql schema."""
-
-    class Meta:
-        model = Section
-
-    code = graphene.String()
-
-    def resolve_code(self, info: graphene.ResolveInfo, **kwargs: Any) -> str:
-        """Resolve the code of the section."""
-        return self.code
-
-
-class _OrganisationObject(MongoengineObjectType):
-    """An organisation in the graphql schema."""
-
-    class Meta:
-        model = Organisation
-
-
-class _OrganisationObjectConnection(graphene.relay.Connection):
-    """A connection for the Organisation object."""
-
-    class Meta:
-        node = _OrganisationObject
-
-
-class _CourseObject(MongoengineObjectType):
-    """A course in the graphql schema."""
-
-    class Meta:
-        model = Course
-
-
-class _CourseObjectConnection(graphene.relay.Connection):
-    """A connection for the Course object."""
-
-    class Meta:
-        node = _CourseObject
-
-
-class _TimetableObject(MongoengineObjectType):
-    """A timetable in the graphql schema."""
-
-    class Meta:
-        model = Timetable
-        exclude_fields = ('key',)
-
-
-class Query(graphene.ObjectType):
-    """A query in the graphql schema."""
-
-    organisation_by_code = graphene.Field(
-        _OrganisationObject, code=graphene.String(required=True)
-    )
-    organisations = graphene.ConnectionField(_OrganisationObjectConnection)
-
-    courses_by_id = graphene.List(
-        _CourseObject, ids=graphene.NonNull(graphene.List(graphene.NonNull(graphene.String))))
-    course_by_id = graphene.Field(
-        _CourseObject, id=graphene.String(required=True))
-    courses = graphene.ConnectionField(_CourseObjectConnection)
-    search_courses = graphene.List(
-        _CourseObject,
-        query=graphene.String(required=True),
-        offset=graphene.Int(required=False, default_value=0),
-        limit=graphene.Int(required=False, default_value=25),
-    )
+class TimetableQuery(graphene.ObjectType):
+    """Timetable queries in the graphql schema."""
 
     timetable_by_id = graphene.Field(
-        _TimetableObject, id=graphene.ID(required=True))
-
-    def resolve_organisation_by_code(
-        self, info: graphene.ResolveInfo, code: str
-    ) -> Organisation:
-        """Return an _OrganisationObject object with the given code."""
-        return Organisation.objects.get(code=code)
-
-    def resolve_organisations(
-        self, info: graphene.ResolveInfo, **kwargs: Any
-    ) -> list[Organisation]:
-        """Return a list of _OrganisationObject objects."""
-        return list(Organisation.objects.all())
-
-    def resolve_courses_by_id(
-            self, info: graphene.ResolveInfo, ids: list[str]) -> list[Course]:
-        """Return a list of _CourseObject objects, each with given ids.
-        Courses that do not exist are null.
-        """
-        courses = []
-        for id in ids:
-            try:
-                course = Course.objects.get(id=id)
-                courses.append(course)
-            except mongoengine.DoesNotExist:
-                courses.append(None)
-        return courses
-
-    def resolve_course_by_id(
-            self, info: graphene.ResolveInfo, id: str) -> Course:
-        """Return a _CourseObject object with the given id."""
-        return Course.objects.get(id=id)
-
-    def resolve_courses(self, info: graphene.ResolveInfo,
-                        **kwargs: Any) -> list[Course]:
-        """Return a list of _CourseObject objects."""
-        return list(Course.objects.all())
-
-    def resolve_search_courses(
-        self, info: graphene.ResolveInfo, query: str, offset: int, limit: int
-    ) -> list[Course]:
-        """Return a list of _CourseObject objects matching the given search string."""
-        # log search query
-        current_app.logger.info(f'searchCourses - query: "{query}", '
-                                f'offset: {offset}, limit: {limit}')
-
-        courses_code = Course.objects(code__icontains=query)
-        # First n search results
-        n = courses_code.count()
-        if offset < n:
-            courses = list(courses_code[offset: offset + limit])
-            if limit > n - offset:
-                courses_text = Course.objects.search_text(
-                    query).order_by('$text_score')
-                courses += list(courses_text.limit(limit - n + offset))
-            return courses
-        else:
-            courses = Course.objects.search_text(query).order_by('$text_score')
-            offset -= n
-            return list(courses[offset: offset + limit])
+        UserTimetableObject, id=graphene.ID(required=True))
 
     def resolve_timetable_by_id(
-            self, _: graphene.ResolveInfo, id: str) -> Timetable:
-        """Return a Timetable object with the given id."""
-        return Timetable.objects.get(id=id)
+        self, info: graphene.ResolveInfo, id: str
+    ) -> Optional[UserTimetable]:
+        """Return a UserTimetable object with the given id."""
+        return UserTimetable.objects.get(id=id)
 
 
 class SetTimetableNameMutation(graphene.Mutation):
@@ -176,7 +30,7 @@ class SetTimetableNameMutation(graphene.Mutation):
         key = graphene.String(required=True)
         name = graphene.String(required=True, default_value=None)
 
-    timetable = graphene.Field(_TimetableObject)
+    timetable = graphene.Field(UserTimetableObject)
     key = graphene.String()
 
     def mutate(
@@ -187,7 +41,7 @@ class SetTimetableNameMutation(graphene.Mutation):
         name: str,
     ) -> 'SetTimetableNameMutation':
         """Set a timetable's name"""
-        timetable = Timetable.objects.get(id=id)
+        timetable = UserTimetable.objects.get(id=id)
         if timetable is None:
             raise Exception(f'could not find timetable with id "{id}"')
 
@@ -206,14 +60,14 @@ class CreateTimetableMutation(graphene.Mutation):
     class Arguments:
         name = graphene.String(required=False, default_value=None)
 
-    timetable = graphene.Field(_TimetableObject)
+    timetable = graphene.Field(UserTimetableObject)
     key = graphene.String()
 
     def mutate(
         self, info: graphene.ResolveInfo, name: Optional[str] = None
     ) -> 'CreateTimetableMutation':
         """Create a timetable."""
-        timetable = Timetable()
+        timetable = UserTimetable()
         if name is not None:
             timetable.name = name
         timetable.save()
@@ -227,18 +81,18 @@ class DuplicateTimetableMutation(graphene.Mutation):
         id = graphene.ID(required=True)
         name = graphene.String(required=False, default_value=None)
 
-    timetable = graphene.Field(_TimetableObject)
+    timetable = graphene.Field(UserTimetableObject)
     key = graphene.String()
 
     def mutate(
             self, info: graphene.ResolveInfo, id: str, name: Optional[str] = None
     ) -> 'DuplicateTimetableMutation':
         """Duplicate a timetable."""
-        timetable = Timetable.objects.get(id=id)
+        timetable = UserTimetable.objects.get(id=id)
         if timetable is None:
             raise Exception(f'could not find timetable with id "{id}"')
 
-        next_timetable = Timetable()
+        next_timetable = UserTimetable()
         if name is not None:
             next_timetable.name = name
 
@@ -257,13 +111,13 @@ class DeleteTimetableMutation(graphene.Mutation):
         id = graphene.ID(required=True)
         key = graphene.String(required=True)
 
-    timetable = graphene.Field(_TimetableObject)
+    timetable = graphene.Field(UserTimetableObject)
 
     def mutate(
         self, _: graphene.ResolveInfo, id: str, key: str
     ) -> 'DeleteTimetableMutation':
         """Delete a timetable."""
-        timetable = Timetable.objects.get(id=id)
+        timetable = UserTimetable.objects.get(id=id)
         if timetable is None:
             raise Exception(f'could not find timetable with id "{id}"')
 
@@ -282,7 +136,7 @@ class RemoveCourseTimetableMutation(graphene.Mutation):
         key = graphene.String(required=True)
         course_id = graphene.String(required=True)
 
-    timetable = graphene.Field(_TimetableObject)
+    timetable = graphene.Field(UserTimetableObject)
 
     def mutate(
         self,
@@ -292,7 +146,7 @@ class RemoveCourseTimetableMutation(graphene.Mutation):
         course_id: str,
     ) -> 'RemoveCourseTimetableMutation':
         """Remove course from a timetable."""
-        timetable = Timetable.objects.get(id=id)
+        timetable = UserTimetable.objects.get(id=id)
         if timetable is None:
             raise Exception(f'could not find timetable with id "{id}"')
 
@@ -317,7 +171,7 @@ class AddSectionsTimetableMutation(graphene.Mutation):
         course_id = graphene.String(required=True)
         sections = graphene.List(graphene.String, required=True)
 
-    timetable = graphene.Field(_TimetableObject)
+    timetable = graphene.Field(UserTimetableObject)
 
     def mutate(
         self,
@@ -328,7 +182,7 @@ class AddSectionsTimetableMutation(graphene.Mutation):
         sections: list[str],
     ) -> 'AddSectionsTimetableMutation':
         """Add sections to a timetable."""
-        timetable = Timetable.objects.get(id=id)
+        timetable = UserTimetable.objects.get(id=id)
         if timetable is None:
             raise Exception(f'could not find timetable with id "{id}"')
 
@@ -362,7 +216,7 @@ class SetSectionsTimetableMutation(graphene.Mutation):
         course_id = graphene.String(required=True)
         sections = graphene.List(graphene.String, required=True)
 
-    timetable = graphene.Field(_TimetableObject)
+    timetable = graphene.Field(UserTimetableObject)
 
     def mutate(
         self,
@@ -376,7 +230,7 @@ class SetSectionsTimetableMutation(graphene.Mutation):
         given course exists in the timetable, then it is removed and replaced
         with the new one.
         """
-        timetable = Timetable.objects.get(id=id)
+        timetable = UserTimetable.objects.get(id=id)
         if timetable is None:
             raise Exception(f'could not find timetable with id "{id}"')
 
@@ -416,8 +270,8 @@ class SetSectionsTimetableMutation(graphene.Mutation):
         return SetSectionsTimetableMutation(timetable=timetable)
 
 
-class Mutation(graphene.ObjectType):
-    """Mutations in the graphql schema."""
+class TimetableMutation(graphene.ObjectType):
+    """Timetable mutations in the graphql schema."""
 
     create_timetable = CreateTimetableMutation.Field()
     set_timetable_name = SetTimetableNameMutation.Field()
@@ -426,6 +280,3 @@ class Mutation(graphene.ObjectType):
     remove_course_timetable = RemoveCourseTimetableMutation.Field()
     add_sections_timetable = AddSectionsTimetableMutation.Field()
     set_sections_timetable = SetSectionsTimetableMutation.Field()
-
-
-schema = graphene.Schema(query=Query, mutation=Mutation)
